requirements.txt
abstract_singleton==1.0.1
beautifulsoup4==4.12.2
deepdiff==6.3.0
gensim==4.3.1
huggingface-hub==0.13.2
networkx==2.8.8
nltk==3.8.1
notebook==6.5.4
node2vec==0.4.6
numpy==1.24.2
openai==1.2.2
pandas==1.5.3
pytesseract==0.3.10
pytest==7.3.1
python-dotenv==1.0.0
regex==2022.10.31
rstr==3.2.1
scipy==1.10.1
seaborn==0.12.2
selenium==4.1.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
tiktoken==0.3.3
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.26.1
webdriver-manager==4.0.0